Add unit test, test selection (#1427)
* add changed script

* remove paths from app.tsx, to eliminate depdendancy link between components and all tests through test utilities

* minor changes to paths exporting

* update casing of urlpathkeytype
