+++ src/admin/adminLayout.component.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Outlet, useLocation } from 'react-router-dom';
 import { BreadcrumbsInfo } from '../api/api.types';
-import { paths } from '../App';
 import BaseLayoutHeader from '../common/baseLayoutHeader.component';
 import PageNotFoundComponent from '../common/pageNotFound/pageNotFound.component';
+import paths from '../paths';
 
 export const AdminErrorComponent = () => {
   return <PageNotFoundComponent homeLocation="Admin" />;
