+++ src/catalogue/catalogueLayout.component.test.tsx
@@ -2,7 +2,7 @@ import { QueryClient } from '@tanstack/react-query';
 import { screen, waitFor } from '@testing-library/react';
 import userEvent, { UserEvent } from '@testing-library/user-event';
 import type { LoaderFunctionArgs } from 'react-router-dom';
-import { paths } from '../App';
+import { URLPathKeyType } from '../paths';
 import { renderComponentWithRouterProvider } from '../testUtils';
 import CatalogueLayout, {
   CatalogueLayoutErrorComponent,
@@ -26,7 +26,7 @@ describe('Catalogue Layout', () => {
   afterEach(() => {
     vi.clearAllMocks();
   });
-  const createView = (path: string, urlPathKey: keyof typeof paths) => {
+  const createView = (path: string, urlPathKey: URLPathKeyType) => {
     return renderComponentWithRouterProvider(
       <CatalogueLayout />,
       urlPathKey,
