+++ src/view/viewTabs.component.tsx
@@ -2,10 +2,10 @@ import Box from '@mui/material/Box';
 import Tabs from '@mui/material/Tabs';
 import React from 'react';
 import { Outlet, useLocation, useNavigate } from 'react-router-dom';
-import { paths } from '../App';
 import { TAB_VALUES, TabValue } from '../app.types';
 import { a11yProps, StyledTab } from '../common/tab/tab.utils';
 import TabPanel from '../common/tab/tabPanel.component';
+import paths from '../paths';
 import { getSciGatewayPageHeightCalc, isRunningInDevelopment } from '../utils';
 
 function ViewTabs() {
